Move callbacks from reactor to structs
