+++ kopf/reactor/activities.py
@@ -19,13 +19,13 @@ The process is intentionally split into multiple packages:
 import logging
 from typing import NoReturn, Mapping
 
-from kopf.reactor import callbacks
 from kopf.reactor import causation
 from kopf.reactor import handlers
 from kopf.reactor import handling
 from kopf.reactor import lifecycles
 from kopf.reactor import registries
 from kopf.reactor import states
+from kopf.structs import callbacks
 from kopf.structs import credentials
 
 logger = logging.getLogger(__name__)
