+++ kopf/reactor/states.py
@@ -55,9 +55,9 @@ import dataclasses
 import datetime
 from typing import Any, Optional, Mapping, Dict, Collection, Iterator, cast, overload
 
-from kopf.reactor import callbacks
 from kopf.reactor import handlers as handlers_
 from kopf.structs import bodies
+from kopf.structs import callbacks
 from kopf.structs import patches
